CosmosDatabases -> CosmosClient CosmosContainers -> CosmosDatabase CosmosItems -> CosmosContainers.Items.cs Scripts -> .\Resource\Scripts Trigger, UDF, StoredProcedures Responses -> Scripts
