+++ Microsoft.Azure.Cosmos/src/Resource/Scripts/CosmosScripts.cs
