+++ Microsoft.Azure.Cosmos/src/Resource/Scripts/CosmosScriptsCore.cs
