+++ Microsoft.Azure.Cosmos/src/Resource/Scripts/ScriptsExtensions.cs
