+++ Microsoft.Azure.Cosmos/src/Resource/Scripts/StoredProcedureResponse.cs
