+++ Microsoft.Azure.Cosmos/src/Resource/Scripts/TriggerResponse.cs
