+++ Microsoft.Azure.Cosmos/src/Resource/Scripts/UserDefinedFunctionResponse.cs
